% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11example3-package.R
\docType{package}
\name{cpp11example3-package}
\alias{cpp11example3}
\alias{cpp11example3-package}
\title{cpp11example3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
